fix: remove deprecation warning from mongo runtime tests
When running the command `yarn test` a below deprecated warning message was printed:

```
(node:75999) DeprecationWarning: current Server Discovery and Monitoring engine is deprecated, and will be removed in a future version. To use the new Server Discover and Monitoring engine, pass option { useUnifiedTopology: true } to the MongoClient constructor.
```

Let's fix the issue by adding the `useUnifiedTopology` options as suggested in the deprecation notice.