test: fix failing runtime-workflow integration tests
Fix the issue by making sure that there is no shared context between each test.

Previous after the first test is executed, subsequent tests were failing with the following message:
```
TypeError: Cannot redefine property: Symbol(apolloServerExecutionDispatcherWillResolveField)

        at Function.defineProperty (<anonymous>)

      at Object.<anonymous> (../node_modules/apollo-server-core/src/requestPipeline.ts:431:14)

      at fulfilled (../node_modules/apollo-server-core/dist/requestPipeline.js:5:58)
 ```
While looking at the code, it seemed to me like the context we passed to apollo server was a global one and hence between multiple tests which resulted in a failure as we attempted to redifine a property that already was defined in the previous ran.