fix(core): getPrimaryKey is undeterministic
If the model contains a primary key field with `@db` annotation and an id scalar field, the computed
primary key will be different depending on the order of declaration. Fix the issue by making sure
that a primary key with `@db` is of high priority while defaulting to an id scalar field.

Fixes https://github.com/aerogear/graphback/issues/1401