fix(codegen-schema): can't add field with custom type to model type
Adding a field with custom type was breaking since the custom type is neither a scalar , enum or a
model type so we cannot really generate the necessary crud bits for it. Let's ignore the field with
undefined type as it is breaking the CRUD bits generation part. `undefined` is returned by this
method
https://github.com/aerogear/graphback/blob/ae5c6d5148877fb7ce6fa622036dcd31ad6b9946/packages/graphback-core/src/crud/mappingHelpers.ts#L159-L173

Fixes https://github.com/aerogear/graphback/issues/1432