fix(schema-plugin): make sure that "null"/"undefined" relationship values are considered resolved
Improve the condition that checked if parent object has the relationship resolved. This also take
care of when the relationship field is "undefined" or "null". With the previous check, we'll query
the database again which seemed wrong. I've added a some data values in the previous tests.

Related to https://github.com/aerogear/graphback/pull/1933 and
https://github.com/aerogear/graphback/pull/1925